requirements.txt
google-generativeai
streamlit
streamlit-chat